refactor(format): extract jsonl export owner
